backend/requirements.txt
fastapi
uvicorn
docling
python-multipart
requests
python-dotenv
openai
google-generativeai
anthropic
